Split stringutil.h into format.h/parse.h; Improve read() API (#301)
