+++ libs/vgc/core/tests/CMakeLists.txt
@@ -1,15 +1,17 @@
 vgc_test_library(core
     CPP_TESTS
         test_array.cpp
+        test_format.cpp
         test_int.cpp
-        test_stringutil.cpp
+        test_parse.cpp
 
     PYTHON_TESTS
         test_color.py
         test_doublearray.py
+        test_format.py
         test_limits.py
         test_object.py
-        test_stringutil.py
+        test_parse.py
         test_vec2d.py
         test_vec2darray.py
 )
