+++ libs/vgc/core/vec2d.h
@@ -21,7 +21,7 @@
 #include <string>
 #include <vgc/core/api.h>
 #include <vgc/core/epsilon.h>
-#include <vgc/core/stringutil.h>
+#include <vgc/core/parse.h>
 
 namespace vgc {
 namespace core {
