+++ libs/vgc/core/vec2darray.cpp
@@ -17,7 +17,7 @@
 #include <vgc/core/vec2darray.h>
 
 #include <sstream>
-#include <vgc/core/stringutil.h>
+#include <vgc/core/parse.h>
 
 namespace vgc {
 namespace core {
